% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_afetada.R
\name{area_afetada}
\alias{area_afetada}
\title{Calcula a percentagem de area afetada}
\usage{
area_afetada(
  im_original,
  im_sadia,
  im_sintoma,
  im_fundo,
  cor_fundo = "branco",
  show_canal = 2,
  salva_image = FALSE,
  dir_original = NULL,
  dir_processada = NULL
)
}
\arguments{
\item{im_original}{A imagem a ser analisada}

\item{im_sadia}{Uma paleta de cores das areas sadias}

\item{im_sintoma}{Uma paleta de cores das areas com sintomas}

\item{im_fundo}{Uma paleta de cores das areas com sintomas}

\item{cor_fundo}{Cor do fundo apos o processamento da imagem}

\item{show_canal}{Qual canal mostrar (1, 2, ou 3)}

\item{salva_image}{Salvar a imagem apos processamento?}

\item{dir_original}{Diretorio que contem as imagens originais}

\item{dir_processada}{Diretorio que contem as imagens processadas}
}
\value{
Um data frame com os resutlados para cada imagem.
}
\description{
Calcula a percentagem de area afetada
}
