% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imagem.R
\name{utils_imagem}
\alias{utils_imagem}
\alias{mostrar_imagem}
\alias{importar_imagem}
\alias{salvar_imagem}
\alias{image_to_mat}
\alias{image_extension}
\alias{image_name}
\title{Importar, exportar e mostrar imagens}
\usage{
mostrar_imagem(imagem, ...)

importar_imagem(imagem, ...)

salvar_imagem(imagem, nome, ...)

image_to_mat(imagem, randomize = TRUE, nrows = 5000)

image_extension(imagem)

image_name(imagem)
}
\arguments{
\item{imagem}{A imagem.
\itemize{
\item Para \code{mostrar_imagem()}, Qualquer objeto R que pode ser coagido a um objeto
raster.
\item Para \code{importar_imagem()}, uma string especificando o caminho para a imagem.
\item Para \code{salvar_imagem()}, um objeto bidimensional ou tridimensional (matriz,
quadro de dados ou array).
}}

\item{...}{Argumentos alternativos a serem passados para outras funcoes:
\itemize{
\item Para \code{mostrar_imagem()}, argumentos passados para \code{\link[grid:grid.raster]{grid::grid.raster()}}.
\item Para \code{importar_imagem()}, argumentos passados para \code{\link[=readPNG]{readPNG()}},
\code{\link[=readJPEG]{readJPEG()}} e \code{\link[=readTIFF]{readTIFF()}}.
\item Para \code{salvar_imagem()}, argumentos passados para \code{\link[=writePNG]{writePNG()}},
\code{\link[=writeJPEG]{writeJPEG()}} e \code{\link[=writeTIFF]{writeTIFF()}}.
}}

\item{nome}{Uma string especificando o nome da imagem a ser salva.}

\item{randomize}{Aleatorizar as linhas apos conversao de matriz para data
frame? Padrao \code{TRUE}.}

\item{nrows}{Numero de linhas a ser selecionada no data frame criado.}
}
\description{
Ferramentas basicas para trabalhar com imagens .png, .jpeg, .jpg, .tiff (ou
.tif, .TIFF, .TIF)
\itemize{
\item \code{importar_imagem()} importa uma imagem.
\item \code{mostrar_imagem()} mostra uma imagem.
\item \code{salvar_imagem()} salva uma imagem.
\item \code{image_to_mat()} converte uma imagem para uma lista, contendo tres
matrizes (RGB) e um data frame com tres colunas (RGB).
\item \code{image_extension()} obtem a extensao de uma imagem.
\item \code{image_name()} obtem o nome de uma imagem.
}
}
\examples{
img1 <- system.file("tmp_images", "img1.png", package = "agrorotinas")
a <- importar_imagem(img1)
mostrar_imagem(a)
}
