% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_area_afetada.R
\name{wrap_area_afetada}
\alias{wrap_area_afetada}
\title{Uma funcao wrapper para a funcao area_afetada}
\usage{
wrap_area_afetada(
  pattern_files,
  im_sadia,
  im_sintoma,
  im_fundo,
  cor_fundo = "branco",
  show_canal = 2,
  salva_image = FALSE,
  dir_original = NULL,
  dir_processada = NULL
)
}
\arguments{
\item{pattern_files}{Um padrao de nome de arquivo}

\item{im_sadia}{Uma paleta de cores das areas sadias}

\item{im_sintoma}{Uma paleta de cores das areas com sintomas}

\item{im_fundo}{Uma paleta de cores das areas com sintomas}

\item{cor_fundo}{Cor do fundo apos o processamento da imagem}

\item{show_canal}{Qual canal mostrar (1, 2, ou 3)}

\item{salva_image}{Salvar a imagem apos processamento?}

\item{dir_original}{Diretorio que contem as imagens originais}

\item{dir_processada}{Diretorio que contem as imagens processadas}
}
\value{
Um data frame
}
\description{
Uma funcao wrapper para a funcao area_afetada
}
