% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_area_afetada.R
\name{wrap_area_afetada}
\alias{wrap_area_afetada}
\title{Uma funcao wrapper para a funcao area_afetada}
\usage{
wrap_area_afetada(pattern_files, im_sadia, im_sintoma, im_fundo, ...)
}
\arguments{
\item{pattern_files}{Um padrao de nome de arquivo}

\item{im_sadia}{Uma paleta de cores das areas sadias}

\item{im_sintoma}{Uma paleta de cores das areas com sintomas}

\item{im_fundo}{Uma paleta de cores das areas com sintomas}

\item{...}{argumentos a serem passados para a funcao \code{\link[=area_afetada]{area_afetada()}}.}
}
\value{
Um data frame
}
\description{
Uma funcao wrapper para a funcao area_afetada
}
